+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEW FEATURES
 
 * Task 2: Dataset censoring component.
+* Task 2: Baseline methods PCA, LMDS and UMAP.
 
 ## MINOR CHANGES